Modifications to allow semi-permanent external data directories to be specified via the pom.xml. Remove need to re-ingest corpora every time evaluation is run.
Squashed commit of the following:

commit f3c8390b7f36bf16b66a04018553f3202ecb8f1c
Author: Matt Pearce (OSC) <matthew.pierce@lexisnexis.co.uk>
Date:   Mon Nov 12 10:05:05 2018 +0000

    Drop forceRefresh parameter from Engine constructor.

commit 815347804646829933db5c8ef4beefcf5a30893c
Author: Matt Pearce (OSC) <matthew.pierce@lexisnexis.co.uk>
Date:   Thu Nov 8 16:59:13 2018 +0000

    Modify ES implementation to use recursive directory deletion.

commit 0d56caf20bc24ca3b07af964ffd0c76ddf10d406
Author: Matt Pearce (OSC) <matthew.pierce@lexisnexis.co.uk>
Date:   Thu Nov 8 16:58:12 2018 +0000

    Implement semi-permanent data directories for Solr.

commit 090088cac8b48e703d977b8006b5194eae155004
Author: Matt Pearce (OSC) <matthew.pierce@lexisnexis.co.uk>
Date:   Wed Nov 7 11:22:58 2018 +0000

    Implement file system checking via Engine class (ES only).

commit f3f8209a59c221802a1e76786903a9a1e23fedbb
Author: Matt Pearce (OSC) <matthew.pierce@lexisnexis.co.uk>
Date:   Tue Nov 6 13:30:56 2018 +0000

    Implement class to check for updates across directories, via checksum file.

+++ rre-search-platform/rre-search-platform-elastic-search-impl/src/main/java/io/sease/rre/search/api/impl/Elasticsearch.java
@@ -3,9 +3,12 @@ package io.sease.rre.search.api.impl;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import io.sease.rre.DirectoryUtils;
 import io.sease.rre.search.api.QueryOrSearchResponse;
 import io.sease.rre.search.api.SearchPlatform;
 import io.sease.rre.search.api.UnableToLoadDataException;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.elasticsearch.action.admin.indices.create.CreateIndexRequest;
 import org.elasticsearch.action.bulk.BulkRequest;
 import org.elasticsearch.action.bulk.BulkResponse;
@@ -25,12 +28,9 @@ import org.elasticsearch.transport.Netty4Plugin;
 
 import java.io.File;
 import java.io.IOException;
-import java.net.URLClassLoader;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
 import java.util.*;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static java.util.Arrays.asList;
 import static java.util.Arrays.stream;
@@ -47,6 +47,8 @@ import static org.elasticsearch.node.InternalSettingsPreparer.prepareEnvironment
  * @since 1.0
  */
 public class Elasticsearch implements SearchPlatform {
+    private static final Logger LOGGER = LogManager.getLogger(Elasticsearch.class);
+
     private static class RRENode extends Node {
         RRENode(final Settings settings, final Collection<Class<? extends Plugin>> plugins) {
             super(prepareEnvironment(settings, null), plugins);
@@ -58,17 +60,28 @@ public class Elasticsearch implements SearchPlatform {
     private final ObjectMapper mapper = new ObjectMapper();
 
     private File nodeConfigFolder;
+    private boolean mustRefresh = false;
 
     @Override
     public void beforeStart(final Map<String, Object> configuration) {
         final File logsFolder = new File("target/elasticsearch/logs");
-        final File dataFolder = new File("target/elasticsearch/data");
+        final File dataFolder = new File((String) configuration.get("path.data"));
 
         logsFolder.delete();
-        dataFolder.delete();
-
         logsFolder.mkdirs();
-        dataFolder.mkdirs();
+
+        if ((Boolean) configuration.get("forceRefresh") && dataFolder.exists()) {
+            try {
+                DirectoryUtils.deleteDirectory(dataFolder);
+            } catch (IOException e) {
+                LOGGER.error("Could not delete data directory - expect data to be stale!", e);
+            }
+            dataFolder.delete();
+        }
+        if (!dataFolder.exists()) {
+            dataFolder.mkdirs();
+            mustRefresh = true;
+        }
 
         nodeConfigFolder = new File((String) configuration.get("path.home"), "config");
         nodeConfigFolder.mkdir();
@@ -171,7 +184,7 @@ public class Elasticsearch implements SearchPlatform {
 
     @Override
     public void beforeStop() {
-        // TODO: remove everything
+        // Nothing to do here, unless we want to delete the data directory
     }
 
     @Override
@@ -202,6 +215,11 @@ public class Elasticsearch implements SearchPlatform {
         } catch (final IOException exception) {
             throw new RuntimeException(exception);
         }
+    }
+
+    @Override
+    public boolean isRefreshRequired() {
+        return mustRefresh;
     }
 
     @SuppressWarnings("unchecked")
